ATOSS/requirements.txt
typing-extensions==4.12.2
sentencepiece==0.1.99
numpy==1.24.3
pandas==1.5.3
pytorch_lightning==1.9.5
tqdm==4.65.0
transformers==4.35.2
openai==0.27.7
Requests==2.31.0
datasets>=2.19.1
accelerate==0.30.1
tyro==0.8.1
trl==0.8.6
jinja2==3.1.3
peft>=0.3.0